chore(build): harden module build plugin and refactor
- Provided source-distribution library lookups are now strict: each requested core/module artifact must match at least one jar, so incomplete lib/provided/* contents fail fast instead of silently falling back to a partial classpath, that is placed in OmegatModuleExtension.groovy.
- Reduce duplication of the definition of a same path among files
- Update to fix the review feedback
